refactor(ui): extract coordinator-admin/components/invites-panel.ts
Move renderInvitesPanel into its own module. The panel now takes a
deps object with { summary, createInvite, renderShell } so the barrel
can wire in the factory action and the shared shell re-render.
